1. Bug Fix: Password isn't compared with current one when checking if password is usable when prompted for a new pasword due to aging. Since a minor version (2.0.3) will occur I added the following changes to be included:
a. Eliminate images not needed due to duplicity creating and maintaining them when the only one used are within the manual itself.
b. As a preventive measure removed Base64 API used from Sun that's proprietary and might be removed to the axis version.

Will perform some test before issuing it to Alex for deployment